Grading
Good work,

In general you need to be a bit more consistent with your commenting throughout the application. As I say in the application the UserSelect filter in the reactive portion should be surrounded by an if() statement.

App works very well minus this error. Lookout for my email with your final grade.